Struct nautilus_indicators::average::dema::DoubleExponentialMovingAverage

#[repr(C)]
pub struct DoubleExponentialMovingAverage {
    pub period: usize,
    pub price_type: PriceType,
    pub value: f64,
    pub count: usize,
    pub is_initialized: bool,
    /* private fields */
}

The Double Exponential Moving Average attempts to a smoother average with less lag than the normal Exponential Moving Average (EMA)

Fields
period: usize

The rolling window period for the indicator (> 0).

price_type: PriceType

The price type used for calculations.

value: f64

The last indicator value.

count: usize

The input count for the indicator.

is_initialized: bool
Implementations

impl DoubleExponentialMovingAverage

pub fn new(period: usize, price_type: Option<PriceType>) -> Result<Self>